package/requirements.txt
tweepy==3.9.0
python-dotenv==0.14.0
requests-oauthlib==1.3.0
python-telegram-bot==13.7
